requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
langchain-core==0.3.15
langgraph==0.2.45
langchain-openai==0.2.5
httpx==0.27.2
python-dotenv==1.0.1
